+++ src/main/java/de/michab/scream/frontend/FrontendX.java
@@ -5,7 +5,7 @@
  * Released under Gnu Public License
  * Copyright (c) 1998-2000 Michael G. Binz
  */
-package de.michab.scream.scanner;
+package de.michab.scream.frontend;
 
 import de.michab.scream.RuntimeX;
